move load openapi from url to init shared lib
