+++ packages/cli/generation/local-generation/local-workspace-runner/package.json
@@ -3,7 +3,7 @@
   "version": "0.0.0",
   "repository": {
     "type": "git",
-    "url": "https://github.com/fern-api/fern.git",
+    "url": "https://github.com/rishan10/fern.git",
     "directory": "packages/cli/generation/local-generation/local-workspace-runner"
   },
   "private": true,
